camp_scout_service/requirements.txt
fastapi[all]==0.89.0
uvicorn[standard]==0.17.6
pytest
psycopg[binary,pool]==3.1.9
jwtdown-fastapi>=0.2.0
python-dotenv
requests==2.30.0
httpx
python-multipart
asyncpg
